pulls useOutsideClickHandler into its own file
